--- a/CHANGELOG.md
+++ b/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 - Add `check_raw_str`, `check_raw_byte_str`, `check_raw_c_str`,
 - Add `unescape_str`, `unescape_byte_str`, `unescape_c_str`,
-- Add `unescape_for_errors`,
+- Add `check_for_errors`,
 - Remove: `unescape_unicode` and `unescape_mixed`
 
 # 0.0.3