--- a/src/lib.rs
+++ b/src/lib.rs
@@ -545,17 +545,25 @@ impl Unescape for CStr {
 /// Enum of the different kinds of literal
 #[derive(Debug, Clone, Copy, PartialEq)]
 pub enum Mode {
+    /// `'a'`
     Char,
 
+    /// `b'a'`
     Byte,
 
+    /// `"hello"`
     Str,
+    /// `r"hello"`
     RawStr,
 
+    /// `b"hello"`
     ByteStr,
+    /// `br"hello"`
     RawByteStr,
 
+    /// `c"hello"`
     CStr,
+    /// `cr"hello"`
     RawCStr,
 }