This improves the API of this crate to not use `unreachable` any more and is the continuation of https://github.com/rust-lang/rust/pull/138163.

It also eliminates internal `unreachable` by inlining `Mode` methods into the *_common functions, and eliminates the resulting duplication by using traits instead. Using traits is much more verbose than a macro-based variant, because they are very explicit, but hopefully also a bit less unclear.

I've tried to use separate commits to explain the story, but have probably only succeeded at the beginning.

There is a companion PR to use this new API here: https://github.com/rust-lang/rust/pull/140999

r? @nnethercote