@nnethercote 2025-05-21
Thanks for splitting this into multiple pieces. I think it's good that @GuillaumeGomez will look at this, it will be good to get another pair of eyes on it after https://github.com/rust-lang/rust/pull/138163.