@GuillaumeGomez 2025-05-21
Changes look good to me. Now comes the not so fun question: can you add benchmarks please?