@hkBst 2025-05-21
> Changes look good to me. Now comes the not so fun question: can you add benchmarks please?

Are you thinking about numbers (before the crate split off this was looking [like this](https://github.com/rust-lang/rust/pull/138163#issuecomment-2720019409) for the macro variant of this change) or code for this crate? I'm happy to come up with some benchmark code.