@GuillaumeGomez 2025-05-21
Mostly code, I can check locally when done. Considering it'll impact performance, better check ahead of time. We just need to check the entry functions.