@hkBst 2025-05-27
Benchmarks PR: https://github.com/rust-lang/literal-escaper/pull/9