@GuillaumeGomez 2025-05-29
Here are the bench results:

| test name | main branch | new changes | diff |
|-|-|-|-|
| bench_check_raw_byte_str | 25,941.68 ns/iter (+/- 57.51) | 48,097.89 ns/iter (+/- 104.59) | +85% |
| bench_check_raw_c_str_ascii | 24,387.77 ns/iter (+/- 83.40) | 39,709.25 ns/iter (+/- 116.80) | +62% |
| bench_check_raw_c_str_unicode | 40,485.60 ns/iter (+/- 181.36) | 50,116.99 ns/iter (+/- 115.94) | +23.7% |
| bench_check_raw_str_ascii | 25,679.40 ns/iter (+/- 68.78) | 39,727.86 ns/iter (+/- 137.46) | +54.7% |
| bench_check_raw_str_unicode | 42,031.95 ns/iter (+/- 199.86) | 47,429.04 ns/iter (+/- 121.53) | +12.8% |
| bench_skip_ascii_whitespace | 6,321.42 ns/iter (+/- 70.54) | 6,322.61 ns/iter (+/- 20.25) | 0% |
| bench_unescape_byte_str_ascii | 64,825.30 ns/iter (+/- 68.90) | 57,407.75 ns/iter (+/- 272.16) | -11.4% |
| bench_unescape_byte_str_hex | 95,602.13 ns/iter (+/- 214.39) | 85,577.19 ns/iter (+/- 188.78) | -10.4% |
| bench_unescape_byte_str_trivial | 34,165.35 ns/iter (+/- 114.53) | 52,296.49 ns/iter (+/- 260.71) | +53% |
| bench_unescape_c_str_ascii | 54,627.06 ns/iter (+/- 76.32) | 48,477.48 ns/iter (+/- 214.53) | -10.7% |
| bench_unescape_c_str_hex_ascii | 94,049.71 ns/iter (+/- 192.47) | 86,601.92 ns/iter (+/- 226.26) | -7.9% |
| bench_unescape_c_str_hex_byte | 94,030.54 ns/iter (+/- 170.87) | 86,285.94 ns/iter (+/- 129.85) | -7.9% |
| bench_unescape_c_str_trivial | 44,069.69 ns/iter (+/- 63.67) | 37,109.07 ns/iter (+/- 89.56) | -15.8% |
| bench_unescape_c_str_unicode | 183,698.16 ns/iter (+/- 236.18) | 165,312.95 ns/iter (+/- 167.56) | -10% |
| bench_unescape_str_ascii | 64,803.76 ns/iter (+/- 88.18) | 66,926.66 ns/iter (+/- 228.30) | +3.3% |
| bench_unescape_str_hex | 95,642.76 ns/iter (+/- 145.34) | 102,020.68 ns/iter (+/- 140.15) | +6.7% |
| bench_unescape_str_trivial | 34,071.43 ns/iter (+/- 54.39) | 46,578.22 ns/iter (+/- 91.29) | +36.7% |
| bench_unescape_str_unicode | 185,521.24 ns/iter (+/- 228.33) | 180,438.70 ns/iter (+/- 428.62) | -2.7% |

Overall, the `check*` functions are much slower. The `unescape*` ones are mixed but with some small gains and big regressions. More work is required to improve everything, we cannot merge as is.