@hkBst 2025-05-30
Interesting! Is there an easy way to create such a nice table?