@GuillaumeGomez 2025-05-30
Sadly no. I ran benches a lot of time in both `main` and in your branch and then kept the lowest `+/-` changes and finally computed the diff for all of them. So I recommend you do the same for `main` and then you can check with your branch.