@hkBst 2025-05-30
Taking the worst offender (`bench_check_raw_byte_str`), if I manually inline `bench_check_raw` (and so get rid of `&mut dyn FnMut`), then a lot (but not all) of the slowdown disappears. Unfortunately just annotating with `#[inline(always)]` does nothing.

On the other hand, if I make the main branch use the newer more generic `bench_check_raw` (which includes adding `+ ?Sized` bounds to `unescape_unicode` and `*_common`), then it becomes just as slow as the new code.

Maybe I should rewrite the benchmarks as macros, to minimize such issues...