@hkBst 2025-05-31
With reduced benchmark overhead commit:

Original (main):
test bench_check_raw_byte_str        ... bench:      25,216.16 ns/iter (+/- 372.96)
test bench_check_raw_c_str_ascii     ... bench:      26,598.25 ns/iter (+/- 113.87)
test bench_check_raw_c_str_unicode   ... bench:      40,294.48 ns/iter (+/- 80.19)
test bench_check_raw_str_ascii       ... bench:      25,423.57 ns/iter (+/- 485.16)
test bench_check_raw_str_unicode     ... bench:      42,309.82 ns/iter (+/- 1,200.37)

Remove Unreachable:
test bench_check_raw_byte_str_ascii  ... bench:      26,611.50 ns/iter (+/- 191.69)  (4% slower)
test bench_check_raw_c_str_ascii     ... bench:      26,445.98 ns/iter (+/- 294.15)  (same)
test bench_check_raw_c_str_unicode   ... bench:      35,513.69 ns/iter (+/- 211.00)  (12% faster)
test bench_check_raw_str_ascii       ... bench:      28,275.46 ns/iter (+/- 78.26) (10% slower)
test bench_check_raw_str_unicode     ... bench:      34,054.50 ns/iter (+/- 133.46) (14% faster)

It is annoying that the ascii path seems to actually be slower, as that is probably the hot path for a lot of code. Will investigate further.

Original (main):
test bench_skip_ascii_whitespace     ... bench:       8,394.66 ns/iter (+/- 27.27)
test bench_unescape_byte_str_ascii   ... bench:      66,282.21 ns/iter (+/- 894.67)
test bench_unescape_byte_str_hex     ... bench:      95,132.18 ns/iter (+/- 1,351.86)
test bench_unescape_byte_str_trivial ... bench:      35,111.39 ns/iter (+/- 782.53)
test bench_unescape_c_str_ascii      ... bench:      55,580.99 ns/iter (+/- 713.06)
test bench_unescape_c_str_hex_ascii  ... bench:      95,237.37 ns/iter (+/- 1,174.20)
test bench_unescape_c_str_hex_byte   ... bench:      95,190.68 ns/iter (+/- 1,317.40)
test bench_unescape_c_str_trivial    ... bench:      46,248.13 ns/iter (+/- 390.16)
test bench_unescape_c_str_unicode    ... bench:     207,910.14 ns/iter (+/- 2,214.58)
test bench_unescape_str_ascii        ... bench:      66,227.90 ns/iter (+/- 870.02)
test bench_unescape_str_hex          ... bench:      95,679.59 ns/iter (+/- 1,811.50)
test bench_unescape_str_trivial      ... bench:      35,406.93 ns/iter (+/- 754.70)
test bench_unescape_str_unicode      ... bench:     207,481.29 ns/iter (+/- 4,115.48)

Remove Unreachable:
test bench_skip_ascii_whitespace     ... bench:       8,404.43 ns/iter (+/- 20.95)  (same)
test bench_unescape_byte_str_ascii   ... bench:      45,606.37 ns/iter (+/- 109.00) (30% faster)
test bench_unescape_byte_str_hex     ... bench:      69,964.92 ns/iter (+/- 143.59) 27% faster)
test bench_unescape_byte_str_trivial ... bench:      31,843.06 ns/iter (+/- 169.52)  (10% faster)
test bench_unescape_c_str_ascii      ... bench:      50,902.69 ns/iter (+/- 968.88)  (10% faster)
test bench_unescape_c_str_hex_ascii  ... bench:      87,675.07 ns/iter (+/- 785.27)  (10% faster)
test bench_unescape_c_str_hex_byte   ... bench:      87,589.62 ns/iter (+/- 564.91)  (10% faster)
test bench_unescape_c_str_trivial    ... bench:      35,969.03 ns/iter (+/- 1,300.87)  (20% faster)
test bench_unescape_c_str_unicode    ... bench:     178,815.79 ns/iter (+/- 1,821.08)  (10% faster)
test bench_unescape_str_ascii        ... bench:      52,682.61 ns/iter (+/- 773.71)  (20% faster)
test bench_unescape_str_hex          ... bench:      90,624.67 ns/iter (+/- 1,252.18)  (5% faster)
test bench_unescape_str_trivial      ... bench:      31,834.04 ns/iter (+/- 580.67)  (10% faster)
test bench_unescape_str_unicode      ... bench:     182,553.44 ns/iter (+/- 3,555.20)  (10% faster)

All of the unescape functions are significantly faster.