@hkBst 2025-06-06
I've switched beck to a while-loop based check_raw to minimize perf difference with previous version.

@GuillaumeGomez can you please check perf again?