@GuillaumeGomez 2025-06-10
Please open a PR with the benchmark changes so I can compare the before/after. Currently I can't compare them since code changed in benchmarks.