@GuillaumeGomez 2025-06-12
Here are the new results:

| bench name | current | new | diff |
|-|-|-|-|
| bench_check_raw_byte_str_ascii | 25366.15 ns/iter (+/- 29.41) | 21362.59 ns/iter (+/- 41.02) | -15.8% |
| bench_check_raw_c_str_ascii | 23746.38 ns/iter (+/- 27.4) | 20248.37 ns/iter (+/- 32.01) | -14.7% |
| bench_check_raw_c_str_non_ascii | 40081.5 ns/iter (+/- 33.36) | 37294.18 ns/iter (+/- 72.35) | -7.0% |
| bench_check_raw_c_str_unicode | 126919.1 ns/iter (+/- 217.34) | 96482.6 ns/iter (+/- 170.2) | -24.0% |
| bench_check_raw_str_ascii | 25425.88 ns/iter (+/- 32.15) | 19805.59 ns/iter (+/- 27.85) | -22.1% |
| bench_check_raw_str_non_ascii | 41483.08 ns/iter (+/- 62.37) | 35902.36 ns/iter (+/- 58.75) | -13.5% |
| bench_check_raw_str_unicode | 134671.66 ns/iter (+/- 134.93) | 109647.88 ns/iter (+/- 132.83) | -18.6% |
| bench_skip_ascii_whitespace | 6327.46 ns/iter (+/- 20.09) | 6305.62 ns/iter (+/- 5.21) | -0.3% |
| bench_unescape_byte_str_ascii | 49183.21 ns/iter (+/- 83.22) | 43504.72 ns/iter (+/- 124.99) | -11.5% |
| bench_unescape_byte_str_ascii_escape | 79799.67 ns/iter (+/- 102.46) | 49577.82 ns/iter (+/- 154.06) | -37.9% |
| bench_unescape_byte_str_hex_escape | 111019.04 ns/iter (+/- 167.05) | 65135.49 ns/iter (+/- 107.09) | -41.3% |
| bench_unescape_byte_str_mixed_escape | 289452.57 ns/iter (+/- 313.3) | 218387.45 ns/iter (+/- 540.27) | -24.6% |
| bench_unescape_c_str_ascii | 64260.92 ns/iter (+/- 175.93) | 54526.39 ns/iter (+/- 67.38) | -15.1% |
| bench_unescape_c_str_ascii_escape | 74902.73 ns/iter (+/- 93.76) | 66932.31 ns/iter (+/- 104.93) | -10.6% |
| bench_unescape_c_str_hex_escape_ascii | 114123.31 ns/iter (+/- 201.84) | 105388.06 ns/iter (+/- 198.93) | -7.7% |
| bench_unescape_c_str_hex_escape_byte | 114360.32 ns/iter (+/- 152.44) | 105959.69 ns/iter (+/- 95.08) | -7.3% |
| bench_unescape_c_str_mixed_escape | 712783.1 ns/iter (+/- 1377.64) | 657509.65 ns/iter (+/- 1073.35) | -7.8% |
| bench_unescape_c_str_non_ascii | 85176.94 ns/iter (+/- 117.64) | 76737.23 ns/iter (+/- 76.32) | -9.9% |
| bench_unescape_c_str_unicode | 297242.9 ns/iter (+/- 571.13) | 262374.38 ns/iter (+/- 349.01) | -11.7% |
| bench_unescape_c_str_unicode_escape | 204612.39 ns/iter (+/- 222.81) | 189687.64 ns/iter (+/- 202.25) | -7.3% |
| bench_unescape_str_ascii | 49467.35 ns/iter (+/- 75.59) | 45187.94 ns/iter (+/- 69.13) | -8.7% |
| bench_unescape_str_ascii_escape | 80234.78 ns/iter (+/- 109.96) | 67375.11 ns/iter (+/- 85.96) | -16.0% |
| bench_unescape_str_hex_escape | 110639.72 ns/iter (+/- 68.65) | 103836.11 ns/iter (+/- 94.97) | -6.1% |
| bench_unescape_str_mixed_escape | 784409.9 ns/iter (+/- 5626.63) | 735422.35 ns/iter (+/- 3384.8) | -6.2% |
| bench_unescape_str_non_ascii | 70945.71 ns/iter (+/- 99.58) | 63450.95 ns/iter (+/- 71.08) | -10.6% |
| bench_unescape_str_unicode | 238028.35 ns/iter (+/- 416.81) | 181429.58 ns/iter (+/- 247.16) | -23.8% |
| bench_unescape_str_unicode_escape | 387049.9 ns/iter (+/- 498.05) | 349056.78 ns/iter (+/- 480.37) | -9.8% |

Only gains, all good. Just for info, I wrote this python script to generate these numbers:

<details><summary>Python script</summary>

```python
import subprocess


class Bench:
    def __init__(self, line):
        self.name = line.split(' ')[1]
        self.speed = float(line.split('... bench:')[1].split('/iter')[0].strip().split(' ')[0].replace(',', ''))
        self.margin = float(line.split('(+/-')[1].split(')')[0].strip().replace(',', ''))


def run_bench():
    res = subprocess.run(["cargo", "+nightly", "bench"], check=True, stdout=subprocess.PIPE)
    return res.stdout.decode('utf-8')


def get_benches():
    print("> Retrieving benches...")
    out = run_bench()
    print("> Done")

    lines = out.split("\n")
    i = 0
    found_bench = False
    while i < len(lines):
        line = lines[i].strip()
        i += 1
        if line.startswith("test result:"):
            found_bench = True
            break
    if not found_bench:
        raise "Not found benchmarks in:\n{}".format(out)
    benches = {}
    while i < len(lines):
        line = lines[i].strip()
        i += 1
        if line.startswith("test result:"):
            break
        if not line.startswith("test bench_"):
            continue
        bench = Bench(line)
        benches[bench.name] = bench
    return benches


def get_good_results():
    benches = get_benches()

    for i in range(0, 30):
        new_benches = get_benches()
        biggest_margin = 0
        updates = 0
        for key in benches.keys():
            if new_benches[key].margin < benches[key].margin:
                benches[key] = new_benches[key]
                updates += 1
            if benches[key].margin > biggest_margin:
                biggest_margin = benches[key].margin
        if biggest_margin > 50.:
            print(">> Biggest margin (iteration {}, updates: {}): {}".format(
                i, updates, biggest_margin))
        else:
            break
    return benches


def get_str_for_bench(bench):
    return "{} ns/iter (+/- {})".format(bench.speed, bench.margin)


def show_results(benches):
    print("==== BENCHES RESULTS =====")
    for value in benches.values():
        print("{}: {}".format(value.name, get_str_for_bench(value)))


def show_comparison(current, new_ones):
    print("=== COMPARISON ===")
    print("| bench name | current | new | diff |")
    print("|-|-|-|-|")
    for key in current.keys():
        print("| {} | {} | {} | {:.1f}% |".format(
            key,
            get_str_for_bench(current[key]),
            get_str_for_bench(new_ones[key]),
            (new_ones[key].speed * 100. / current[key].speed) - 100.,
        ))


def main():
    subprocess.run(["git", "checkout", "hkBst/benches"], check=True)
    current = get_good_results()
    show_results(current)
    subprocess.run(["git", "checkout", "hkBst/remove_unreachable"], check=True)
    new_ones = get_good_results()
    show_results(new_ones)
    show_comparison(current, new_ones)


if __name__ == "__main__":
    main()
```

</details>